+++ packages/ndla-editor/src/FileStructure.js
@@ -9,62 +9,14 @@
 import React, { Component, Fragment } from 'react';
 import PropTypes from 'prop-types';
 import styled, { css } from 'react-emotion';
-import { spacing, colors, fonts } from '@ndla/core';
+import { spacing, colors } from '@ndla/core';
 import Spinner from './Spinner';
+import ItemName from './ItemName';
 
 const ItemListWrapper = styled.ul`
   margin: 0;
   padding: 0;
 `;
-
-const itemTitleArrow = css`
-  &:before {
-    content: '';
-    display: block;
-    width: 0;
-    height: 0;
-    border-top: 6px solid transparent;
-    border-bottom: 6px solid transparent;
-    border-left: 9px solid ${colors.text.primary};
-    margin-right: ${spacing.xsmall};
-  }
-`;
-
-const itemTitleLinked = css`
-  &:before {
-    content: '';
-    display: block;
-    width: 8px;
-    height: 8px;
-    border-bottom: 2px solid ${colors.brand.light};
-    border-left: 2px solid ${colors.brand.light};
-    border-bottom-left-radius: 2px;
-    margin-right: ${spacing.xsmall};
-    margin-left: 7px;
-  }
-`;
-
-const ItemTitleButton = styled.button`
-  ${fonts.sizes(16, 1)} font-weight: ${fonts.weight.semibold};
-  border: 0;
-  background: 0;
-  color: ${colors.brand.primary};
-  display: flex;
-  align-items: center;
-  text-align: left;
-  white-space: nowrap;
-  ${props => props.arrowDirection !== undefined && itemTitleArrow};
-  ${props =>
-    props.arrowDirection === undefined &&
-    !props.isMainTopic &&
-    itemTitleLinked};
-  &:before {
-    transition: transform 200ms ease;
-    transform: rotate(${props => props.arrowDirection}deg);
-  }
-`;
-
-const ItemTitleSpan = ItemTitleButton.withComponent('span');
 
 const ItemsList = styled.li`
   margin: 0;
@@ -97,45 +49,6 @@ const ItemsList = styled.li`
     `};
 `;
 
-const itemNameStyling = css`
-  display: flex;
-  justify-content: space-between;
-`;
-
-const ItemName = ({
-  title,
-  children,
-  path,
-  toggleOpen,
-  hasSubtopics,
-  isOpen,
-  level,
-}) => (
-  <div className={itemNameStyling}>
-    {hasSubtopics ? (
-      <ItemTitleButton
-        type="button"
-        arrowDirection={isOpen ? 90 : 0}
-        onClick={() => toggleOpen(path)}>
-        {title}
-      </ItemTitleButton>
-    ) : (
-      <ItemTitleSpan isMainTopic={level === 0}>{title}</ItemTitleSpan>
-    )}
-    {children}
-  </div>
-);
-
-ItemName.propTypes = {
-  title: PropTypes.string.isRequired,
-  children: PropTypes.node,
-  path: PropTypes.string.isRequired,
-  toggleOpen: PropTypes.func.isRequired,
-  hasSubtopics: PropTypes.bool.isRequired,
-  isOpen: PropTypes.bool,
-  level: PropTypes.number,
-};
-
 class FileStructure extends Component {
   constructor(props) {
     super(props);
@@ -151,10 +64,7 @@ class FileStructure extends Component {
       if (index === -1) {
         // Has other subjects open and !allowMultipleSubjectsOpen?
         if (level === 0 && !this.props.allowMultipleSubjectsOpen) {
-          openedPaths = openedPaths.filter(
-            openPath =>
-              !this.props.structure.some(subject => subject.id === openPath),
-          );
+          openedPaths = [];
         }
         openedPaths.push(path);
         if (this.props.onOpenPath) {
@@ -175,14 +85,14 @@ class FileStructure extends Component {
     const {
       renderListItems,
       listClass,
-      FileStructureFilters,
+      fileStructureFilters,
       filters,
     } = this.props;
 
     const ignoreFilter =
       level === 0 ||
       !filters[subjectId].some(filter =>
-        FileStructureFilters.includes(filter.id),
+        fileStructureFilters.includes(filter.id),
       );
 
     return (
@@ -192,7 +102,7 @@ class FileStructure extends Component {
             topic =>
               ignoreFilter ||
               topic.filters.some(topicFilter =>
-                FileStructureFilters.includes(topicFilter.id),
+                fileStructureFilters.includes(topicFilter.id),
               ),
           )
           .map(topic => {
@@ -258,7 +168,7 @@ FileStructure.propTypes = {
   renderListItems: PropTypes.func,
   listClass: PropTypes.string,
   onOpenPath: PropTypes.func,
-  FileStructureFilters: PropTypes.arrayOf(PropTypes.string),
+  fileStructureFilters: PropTypes.arrayOf(PropTypes.string),
   filters: PropTypes.shape(
     PropTypes.arrayOf(
       PropTypes.shape({
@@ -273,7 +183,7 @@ FileStructure.defaultProps = {
   structure: [],
   openedPaths: [],
   className: '',
-  FileStructureFilters: [],
+  fileStructureFilters: [],
   filters: [],
   allowMultipleSubjectsOpen: false,
 };
